Put stuff into place to write out economic data json

+++ src/EconomyDataLoader/EconomyDataLoader/Models/Output/BlsDataValues.cs
@@ -3,7 +3,7 @@
 public class BlsDataValues
 {
     [JsonPropertyName("unemploymentData")]
-    public int? UnemploymentData { get; set; } = null;
+    public float? UnemploymentData { get; set; } = null;
 
     [JsonPropertyName("cpi")]
     public float? Cpi { get; set; } = null;
